requirements.txt
python-telegram-bot[asyncio]
aiohttp
python-dotenv